+++ lib_dsp/api/lib_dsp.h
@@ -1,15 +1,17 @@
-// Copyright (c) 2015, XMOS Ltd, All rights reserved
+// Copyright (c) 2016, XMOS Ltd, All rights reserved
 
 #ifndef __dsp_h__
 #define __dsp_h__
 
 #include <lib_dsp_adaptive.h>
+#include <lib_dsp_design.h>
 #include <lib_dsp_filters.h>
 #include <lib_dsp_matrix.h>
 #include <lib_dsp_statistics.h>
 #include <lib_dsp_math.h>
 #include <lib_dsp_qformat.h>
 #include <lib_dsp_vector.h>
+#include <lib_dsp_transforms.h>
 
 #endif